# service/identity/internal/httpapi/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Beexster Identity Service API
  version: 1.0.0
  description: |
    Identity lifecycle, email/password authentication, link-based verification,
    password recovery, refresh-token rotation, session management, platform
    administrative roles, soft deletion, reactivation and login history.
servers:
  - url: http://localhost:8080
security: []
...
paths:
...
  /v1/auth/logout:
    post:
      tags: [Sessions]
      operationId: logoutCurrent
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Current refresh session revoked.}
        '401': {$ref: '#/components/responses/InvalidToken'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/auth/logout-all:
    post:
      tags: [Sessions]
      operationId: logoutAll
      security: [{bearerAuth: []}]
      responses:
        '204': {description: All refresh sessions revoked.}
        '401': {$ref: '#/components/responses/InvalidToken'}
        '500': {$ref: '#/components/responses/InternalError'}
...
  /v1/me/login-history:
    get:
      tags: [Identity]
      operationId: loginHistory
      security: [{bearerAuth: []}]
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 25}
        - name: before
          in: query
          description: RFC3339 cursor from a prior response.
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: Login attempts associated with the current identity.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginHistoryResponse'}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '401': {$ref: '#/components/responses/InvalidToken'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    InvalidInput:
      description: Invalid request input.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {error: {code: ERR_INVALID_INPUT, message: invalid input, request_id: 0123456789abcdef}}
...
    InvalidToken:
      description: Invalid, expired, revoked or reused token.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    InternalError:
      description: Internal dependency failure. Sensitive operations fail closed if Redis is unavailable.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
    UUIDv7:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - ERR_INVALID_INPUT
                - ERR_EMAIL_ALREADY_EXISTS
                - ERR_INVALID_CREDENTIALS
                - ERR_EMAIL_NOT_VERIFIED
                - ERR_ACCOUNT_INACTIVE
                - ERR_TOKEN_INVALID
                - ERR_TOKEN_EXPIRED
                - ERR_TOKEN_REUSE_DETECTED
                - ERR_NOT_FOUND
                - ERR_FORBIDDEN
                - ERR_RATE_LIMITED
                - ERR_INTERNAL
            message: {type: string}
            request_id: {type: string}
...
    LoginHistoryResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [attempts]
          properties:
            attempts:
              type: array
              items:
                type: object
                required: [id, success, ip_address, user_agent, attempted_at]
                properties:
                  id: {$ref: '#/components/schemas/UUIDv7'}
                  success: {type: boolean}
                  failure_code: {type: string}
                  ip_address: {type: string}
                  user_agent: {type: string}
                  attempted_at: {type: string, format: date-time}
            next_before: {type: string, format: date-time}
...